feat(): electron builder
So far I haven't seen the best way to build an app with electron and python. The only way to generate a functional build was to generate an app.asar with electron-builder and the python executable with pyinstaller. Then it becomes just a wrapper to start the server with eel and use electron as a browser and with a compiled code in the form of asar file.